CR-CHANGES /FormFields cmp, constants, file names, hash lib
