+++ frontend/src/constants/localStorageKeys.js
@@ -1,1 +1,1 @@
-export const USER_STORAGE_KEY = "registered_users";
+export const USER_STORAGE_KEY = "registered_user";
